Refactor specification package
