+++ final_task/pycalc/specification/__init__.py
@@ -1,0 +1,5 @@
+"""
+Specification for a parser.
+"""
+
+from .specification import Specification
